[#152] refactor: WorkTime 경로 리팩터링
