--- server/src/main/java/com/gigker/server/domain/common/WorkTimeResponseDto.java
@@ -1,13 +1,0 @@
-package com.gigker.server.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-public class WorkTimeResponseDto {
-    private LocalDateTime startWorkTime;
-    private LocalDateTime endWorkTime;
-}
