+++ server/src/main/java/com/gigker/server/domain/schedule/dto/ScheduleResponseDto.java
@@ -3,7 +3,7 @@ package com.gigker.server.domain.schedule.dto;
 import java.util.List;
 
 import com.gigker.server.domain.common.ContentType;
-import com.gigker.server.domain.common.WorkTimeResponseDto;
+import com.gigker.server.domain.workTime.dto.WorkTimeResponseDto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,10 +13,10 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class ScheduleResponseDto {
-    private Long contentId;
-    private String title;
-    private ContentType contentType;
-    private Integer recruitingCount;
-    private List<WorkTimeResponseDto> workTimes;
-    private int price;
+	private Long contentId;
+	private String title;
+	private ContentType contentType;
+	private Integer recruitingCount;
+	private List<WorkTimeResponseDto> workTimes;
+	private int price;
 }
